refactor(webapp): regroup budget-line dialogs under shared dir
Merge create-budget-line/ + edit-budget-line/ into budget-line/{create,edit}/.
Strip redundant entity prefix from filenames (dialog.ts, dialog.schema.ts) —
class names, selectors, and data-testid attributes unchanged.

Mirrors the recent allocated-transactions refactor (b6a84b1ab).
