+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-line/create/dialog.schema.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 import {
   budgetLineCreateFromFormSchema,
   type BudgetLineCreateFormValue,
-} from './add-budget-line-dialog.schema';
+} from './dialog.schema';
 
 const BUDGET_ID = '00000000-0000-4000-8000-000000000001';
