+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-line/edit/dialog.ts
@@ -1,25 +1,30 @@
 import {
-  ChangeDetectionStrategy,
   Component,
-  computed,
   inject,
+  ChangeDetectionStrategy,
+  computed,
   signal,
 } from '@angular/core';
-import { Field, form, minLength, required } from '@angular/forms/signals';
 import {
+  MatDialogRef,
   MAT_DIALOG_DATA,
   MatDialogModule,
-  MatDialogRef,
 } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { Field, form, required, minLength } from '@angular/forms/signals';
+import {
+  type BudgetLine,
+  type BudgetLineUpdate,
+  type TransactionKind,
+  type TransactionRecurrence,
+} from 'pulpe-shared';
 import { TranslocoPipe } from '@jsverse/transloco';
-import { type TransactionKind, type TransactionRecurrence } from 'pulpe-shared';
-
+import { TransactionIconPipe } from '@ui/transaction-display';
+import { TransactionLabelPipe } from '@ui/transaction-display';
 import {
   applyAmountValidators,
   type AmountFormSlice,
@@ -27,29 +32,24 @@ import {
   CurrencyConverterService,
 } from '@core/currency';
 import { UserSettingsStore } from '@core/user-settings';
-import { touchedFieldErrors } from '@core/validators';
+import { FeatureFlagsService } from '@core/feature-flags';
+import { touchedFieldErrors } from '@core/validators/touched-field-errors';
 import { AmountInput } from '@app/pattern/amount-input/amount-input';
-import {
-  TransactionIconPipe,
-  TransactionLabelPipe,
-} from '@ui/transaction-display';
+import { budgetLineUpdateFromFormSchema } from './dialog.schema';
 
-import { budgetLineCreateFromFormSchema } from './add-budget-line-dialog.schema';
-
-export interface BudgetLineDialogData {
-  budgetId: string;
+export interface EditBudgetLineDialogData {
+  budgetLine: BudgetLine;
 }
 
-interface AddBudgetLineModel {
+interface EditBudgetLineModel {
   name: string;
+  money: AmountFormSlice;
   kind: TransactionKind;
   recurrence: TransactionRecurrence;
-  isChecked: boolean;
-  money: AmountFormSlice;
 }
 
 @Component({
-  selector: 'pulpe-budget-line-dialog',
+  selector: 'pulpe-edit-budget-line-dialog',
   imports: [
     MatDialogModule,
     MatFormFieldModule,
@@ -57,17 +57,16 @@ interface AddBudgetLineModel {
     MatSelectModule,
     MatButtonModule,
     MatIconModule,
-    MatSlideToggleModule,
     TranslocoPipe,
     TransactionIconPipe,
     TransactionLabelPipe,
     Field,
     AmountInput,
   ],
-  host: { 'data-testid': 'add-budget-line-dialog' },
+  host: { 'data-testid': 'edit-budget-line-dialog' },
   template: `
     <h2 mat-dialog-title class="text-headline-small">
-      {{ 'budget.newForecast' | transloco }}
+      {{ 'budget.editForecast' | transloco }}
     </h2>
 
     <mat-dialog-content>
@@ -81,9 +80,9 @@ interface AddBudgetLineModel {
             <mat-label>{{ 'budget.forecastNameLabel' | transloco }}</mat-label>
             <input
               matInput
-              [field]="addForm.name"
+              [field]="editForm.name"
               [placeholder]="'budget.forecastNamePlaceholder' | transloco"
-              data-testid="new-line-name"
+              data-testid="edit-line-name"
             />
             @if (nameErrors().required) {
               <mat-error>{{
@@ -97,7 +96,11 @@ interface AddBudgetLineModel {
             }
           </mat-form-field>
 
-          <pulpe-amount-input [control]="addForm.money" />
+          <pulpe-amount-input
+            [control]="editForm.money"
+            mode="edit"
+            [originalCurrency]="originalCurrency"
+          />
 
           <mat-form-field
             appearance="outline"
@@ -105,7 +108,7 @@ interface AddBudgetLineModel {
             class="w-full"
           >
             <mat-label>{{ 'budget.forecastTypeLabel' | transloco }}</mat-label>
-            <mat-select [field]="addForm.kind" data-testid="new-line-kind">
+            <mat-select [field]="editForm.kind" data-testid="edit-line-kind">
               <mat-option value="income">
                 <mat-icon class="text-financial-income">{{
                   'income' | transactionIcon
@@ -131,85 +134,97 @@ interface AddBudgetLineModel {
               }}</mat-error>
             }
           </mat-form-field>
-
-          <div class="flex items-center justify-between py-2 px-1">
-            <span class="text-body-medium text-on-surface">{{
-              'budget.forecastCheckedToggle' | transloco
-            }}</span>
-            <mat-slide-toggle
-              [field]="addForm.isChecked"
-              [attr.aria-label]="'budget.forecastCheckedToggle' | transloco"
-            />
-          </div>
         </div>
       </div>
     </mat-dialog-content>
 
     @if (conversionError()) {
-      <p class="text-error text-body-small px-6 pb-2">
+      <p role="alert" class="text-error text-body-small px-6 pb-2">
         {{ 'common.conversionError' | transloco }}
       </p>
     }
     <mat-dialog-actions align="end">
-      <button matButton (click)="cancel()" data-testid="cancel-new-line">
+      <button matButton (click)="handleCancel()" data-testid="cancel-edit-line">
         {{ 'common.cancel' | transloco }}
       </button>
       <button
         matButton="filled"
         color="primary"
-        (click)="submit()"
+        (click)="handleSubmit()"
         [disabled]="!canSubmit()"
-        data-testid="add-new-line"
+        data-testid="save-edit-line"
       >
-        <mat-icon>add</mat-icon>
-        {{ 'common.add' | transloco }}
+        <mat-icon>save</mat-icon>
+        {{ 'common.save' | transloco }}
       </button>
     </mat-dialog-actions>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class AddBudgetLineDialog {
-  readonly #dialogRef = inject(MatDialogRef<AddBudgetLineDialog>);
-  readonly #data = inject<BudgetLineDialogData>(MAT_DIALOG_DATA);
+export class EditBudgetLineDialog {
+  readonly #dialogRef = inject(MatDialogRef<EditBudgetLineDialog>);
+  readonly #data = inject<EditBudgetLineDialogData>(MAT_DIALOG_DATA);
   readonly #settings = inject(UserSettingsStore);
+  readonly #flags = inject(FeatureFlagsService);
   readonly #converter = inject(CurrencyConverterService);
 
-  protected readonly model = signal<AddBudgetLineModel>({
-    name: '',
-    kind: 'expense',
-    recurrence: 'one_off',
-    isChecked: false,
-    money: createAmountSlice({ initialCurrency: this.#settings.currency() }),
+  protected readonly originalCurrency =
+    this.#data.budgetLine.originalCurrency ?? this.#settings.currency();
+
+  protected readonly showCurrencySelector = computed(
+    () =>
+      this.#flags.isMultiCurrencyEnabled() &&
+      this.originalCurrency !== this.#settings.currency(),
+  );
+
+  protected readonly model = signal<EditBudgetLineModel>({
+    name: this.#data.budgetLine.name,
+    money: createAmountSlice({
+      initialCurrency: this.originalCurrency,
+      initialAmount: this.#computeInitialAmount(),
+    }),
+    kind: this.#data.budgetLine.kind,
+    recurrence: this.#data.budgetLine.recurrence,
   });
 
-  protected readonly addForm = form(this.model, (path) => {
+  protected readonly editForm = form(this.model, (path) => {
     required(path.name, { message: 'budget.forecastNameRequired' });
     minLength(path.name, 1);
     applyAmountValidators(path.money);
     required(path.kind, { message: 'budget.forecastTypeRequired' });
+    required(path.recurrence);
   });
-
-  protected readonly nameErrors = touchedFieldErrors(
-    () => this.addForm.name,
-    'required',
-    'minLength',
-  );
-  protected readonly kindErrors = touchedFieldErrors(
-    () => this.addForm.kind,
-    'required',
-  );
 
   protected readonly conversionError = signal(false);
   protected readonly isSubmitting = signal(false);
   protected readonly canSubmit = computed(
-    () => this.addForm().valid() && !this.isSubmitting(),
+    () => this.editForm().valid() && !this.isSubmitting(),
   );
 
-  protected async submit(): Promise<void> {
-    if (!this.canSubmit()) return;
+  protected readonly nameErrors = touchedFieldErrors(
+    () => this.editForm.name,
+    'required',
+    'minLength',
+  );
+  protected readonly kindErrors = touchedFieldErrors(
+    () => this.editForm.kind,
+    'required',
+  );
+
+  #computeInitialAmount(): number {
+    const line = this.#data.budgetLine;
+    if (this.showCurrencySelector() && line.originalAmount != null) {
+      return line.originalAmount;
+    }
+    return line.amount;
+  }
+
+  async handleSubmit(): Promise<void> {
+    if (!this.editForm().valid()) return;
 
     this.conversionError.set(false);
     this.isSubmitting.set(true);
+
     try {
       const m = this.model();
       const { convertedAmount, metadata } =
@@ -219,16 +234,20 @@ export class AddBudgetLineDialog {
           this.#settings.currency(),
         );
 
-      const dto = budgetLineCreateFromFormSchema.parse({
-        budgetId: this.#data.budgetId,
-        name: m.name.trim(),
+      const formPart = budgetLineUpdateFromFormSchema.parse({
+        name: m.name,
         amount: convertedAmount,
         kind: m.kind,
         recurrence: m.recurrence,
-        isChecked: m.isChecked,
         conversion: metadata,
       });
-      this.#dialogRef.close(dto);
+      const update: BudgetLineUpdate = {
+        id: this.#data.budgetLine.id,
+        templateLineId: this.#data.budgetLine.templateLineId,
+        savingsGoalId: this.#data.budgetLine.savingsGoalId,
+        ...formPart,
+      };
+      this.#dialogRef.close(update);
     } catch {
       this.conversionError.set(true);
     } finally {
@@ -236,7 +255,7 @@ export class AddBudgetLineDialog {
     }
   }
 
-  protected cancel(): void {
+  handleCancel(): void {
     this.#dialogRef.close();
   }
 }
